backend/requirements-dev.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic-settings>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
transformers>=4.36.0
torch>=2.0.0
torchaudio>=2.0.0
faster-whisper>=1.1.1
opencv-python-headless>=4.8.0
Pillow>=10.0.0
ffmpeg-python>=0.2.0
nltk>=3.8.0
pandas>=2.0.0
tqdm>=4.65.0
spacy>=3.7.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
networkx>=3.1
pyyaml>=6.0
requests>=2.31.0
aiofiles>=23.0.0
google-generativeai>=0.8.0
PyJWT>=2.8.0
bcrypt>=4.0.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
